+++ src/redturtle/volto/tests/test_service_sitemap.py
@@ -1,0 +1,51 @@
+# -*- coding: utf-8 -*-
+from plone import api
+from plone.app.testing import (
+    SITE_OWNER_NAME,
+    SITE_OWNER_PASSWORD,
+    TEST_USER_ID,
+    setRoles,
+)
+from plone.restapi.testing import RelativeSession
+from redturtle.volto.testing import REDTURTLE_VOLTO_API_FUNCTIONAL_TESTING
+from transaction import commit
+
+import unittest
+
+
+class TestServicesSitemap(unittest.TestCase):
+
+    layer = REDTURTLE_VOLTO_API_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        self.app = self.layer["app"]
+        self.portal = self.layer["portal"]
+        self.portal_url = self.portal.absolute_url()
+        setRoles(self.portal, TEST_USER_ID, ["Manager"])
+
+        self.api_session = RelativeSession(self.portal_url)
+        self.api_session.headers.update({"Accept": "application/json"})
+        self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
+
+    def tearDown(self):
+        self.api_session.close()
+
+    def test_route_exists(self):
+        response = self.api_session.get("/@sitemap-settings")
+
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.headers.get("Content-Type"), "application/json")
+
+    def test_return_values_from_registry(self):
+        response = self.api_session.get("/@sitemap-settings")
+        result = response.json()
+
+        self.assertEqual(result["depth"], 3)
+
+        api.portal.set_registry_record("plone.sitemap_depth", 5)
+        commit()
+
+        response = self.api_session.get("/@sitemap-settings")
+        result = response.json()
+
+        self.assertEqual(result["depth"], 5)